+++ .gitignore
@@ -3,7 +3,7 @@ node_modules
 # Output
 .output
 .vercel
-/.svelte-kit
+.next/
 /build
 
 # OS
@@ -16,12 +16,5 @@ Thumbs.db
 !.env.example
 !.env.test
 
-# Vite
-vite.config.js.timestamp-*
-vite.config.ts.timestamp-*
-react-try/node_modules
-react-try/dist
-react-try/tsconfig*.tsbuildinfo
-react-try/vite.config.js
-react-try/vite.config.d.ts
+# Build info
 tsconfig.tsbuildinfo
