+++ package.json
@@ -1,12 +1,11 @@
 {
-  "name": "react-try",
+  "name": "captive-portal",
   "version": "0.0.1",
   "private": true,
-  "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc -b && vite build",
-    "preview": "vite preview"
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start"
   },
   "dependencies": {
     "@hookform/resolvers": "^5.2.2",
@@ -38,8 +37,6 @@
     "@radix-ui/react-toggle-group": "^1.1.11",
     "@radix-ui/react-tooltip": "^1.2.8",
     "@tailwindcss/postcss": "^4.1.13",
-    "@tailwindcss/vite": "^4.1.13",
-    "@vitejs/plugin-react": "^5.0.4",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
@@ -47,6 +44,7 @@
     "embla-carousel-react": "^8.6.0",
     "input-otp": "^1.4.2",
     "lucide-react": "^0.544.0",
+    "next": "^15.0.3",
     "next-themes": "^0.4.6",
     "postcss": "^8.5.6",
     "react": "^19.1.1",
@@ -54,7 +52,6 @@
     "react-dom": "^19.1.1",
     "react-hook-form": "^7.63.0",
     "react-resizable-panels": "^3.0.6",
-    "react-router-dom": "^7.9.3",
     "recharts": "2.15.4",
     "sonner": "^2.0.7",
     "tailwind-merge": "^3.3.1",
@@ -66,11 +63,9 @@
     "@types/node": "^24.6.0",
     "@types/react": "^19.1.16",
     "@types/react-dom": "^19.1.9",
-    "@vitejs/plugin-react": "^5.0.4",
     "autoprefixer": "^10.4.20",
     "tailwindcss": "^4.1.13",
     "tw-animate-css": "^1.4.0",
-    "typescript": "^5.6.3",
-    "vite": "^7.1.7"
+    "typescript": "^5.6.3"
   }
 }
