--- src/main.tsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import { BrowserRouter } from 'react-router-dom';
-
-import App from './App';
-import './index.css';
-import { TranslationProvider } from './lib/i18n';
-
-ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
-  <React.StrictMode>
-    <BrowserRouter>
-      <TranslationProvider>
-        <App />
-      </TranslationProvider>
-    </BrowserRouter>
-  </React.StrictMode>
-);
